+++ doc/api_example.py
@@ -20,7 +20,7 @@ bad_exchanges = []
 for exch in exchanges:
     exch.silence([1089, 1194])      # Errors we don't care about
     httpolice.check_exchange(exch)
-    if any(notice.severity == httpolice.ERROR
+    if any(notice.severity > httpolice.Severity.comment
            for resp in exch.responses       # We only care about responses
            for notice in resp.notices):
         bad_exchanges.append(exch)