+++ httpolice/__init__.py
@@ -3,16 +3,22 @@
 from httpolice import helpers
 from httpolice.__metadata__ import version as __version__
 from httpolice.exchange import Exchange, check_exchange
-from httpolice.notice import COMMENT, DEBUG, ERROR
+from httpolice.notice import Severity
 from httpolice.reports.html import html_report
 from httpolice.reports.text import text_report
 from httpolice.request import Request
 from httpolice.response import Response
 
+# Backward compatibility
+ERROR = Severity.error
+COMMENT = Severity.comment
+DEBUG = Severity.debug
+
 __all__ = [
     'Exchange',
     'Request',
     'Response',
+    'Severity',
     'COMMENT',
     'DEBUG',
     'ERROR',